backend/requirements.txt
fastapi
pydantic
gunicorn
uvicorn
python-dotenv
requests
python-multipart